Convert to modern csproj format
Motivation
----------
A key step towards supporting .NET Core, as well as making for a
cleaner and easier to manage project file.

Modifications
-------------
Import the .NET SDK at the top of the project file, and drop the
manual imports and properties that are no longer necessary.

Drop all files references from the project files, globbing is now
automatic for most of them.

Convert from packages.config to PackageReference in the project file.

Drop AssemblyInfo.cs files, these attributes are now generated from
properties in the project file.

Include NuGet package properties in the project files to support
automatic packing.

Add Aspect.targets file to include it in the package.