[WIP] Correct try/catch leave behavior
Motivation
----------
When adding a try/catch block, the current approach of leaving ret
operations within the try block is semantically incorrect for MSIL.

Modifications
-------------
Rework the try/catch block to rewrite the ret instructions as leave
instructions, adding a new ret at the end of the method outside the
try/catch block. Save the return value, if any, and reload outside the
try/catch.

Move the call to InsertOnExitCalls after the OnException handling so
that it only inserts calls before the new, final ret instruction.

Add ReplaceInstruction and ReplaceInstructionAndLeaveTarget helpers to
IlGenerator, and rework InsertOnExitCalls to also use these helpers.
This manages the replacement of branch statements recursively, and also
adjusts exception blocks.